+++ bindings/nodejs/examples/how_tos/simple_transaction/simple-transaction.ts
@@ -35,7 +35,7 @@ async function run() {
             'rms1qrrv7flg6lz5cssvzv2lsdt8c673khad060l4quev6q09tkm9mgtupgf0h0';
         const amount = '1000000';
 
-        const response = await account.sendAmount([
+        const response = await account.send([
             {
                 address,
                 amount,