+++ bindings/python/examples/how_tos/native_tokens/create.py
@@ -1,6 +1,5 @@
 from iota_sdk import Wallet, utf8_to_hex
 from dotenv import load_dotenv
-import time
 import os
 
 load_dotenv()
@@ -17,16 +16,16 @@ if 'STRONGHOLD_PASSWORD' not in os.environ:
 wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 
 # Sync account with the node
-response = account.sync()
+balance = account.sync()
 
 # We can first check if we already have an alias in our account, because an alias can have many foundry outputs and therefore we can reuse an existing one
-if len(account.aliases) == 0:
+if not balance["aliases"]:
     # If we don't have an alias, we need to create one
     transaction = account.prepare_create_alias_output(None, None).send()
     print(f'Transaction sent: {transaction["transactionId"]}')
 
     # Wait for transaction to get included
-    blockId = account.retry_transaction_until_included(transaction['transactionId'])
+    blockId = account.retry_transaction_until_included(transaction.transactionId)
     print(f'Block included: {os.environ["EXPLORER_URL"]}/block/{blockId}')
 
     account.sync()
@@ -42,10 +41,10 @@ params = {
 
 prepared_transaction = account.prepare_create_native_token(params, None)
 transaction = prepared_transaction.send()
-print(f'Transaction sent: {transaction["transactionId"]}')
+print(f'Transaction sent: {transaction.transactionId}')
 
 # Wait for transaction to get included
-blockId = account.retry_transaction_until_included(transaction['transactionId'])
+blockId = account.retry_transaction_until_included(transaction.transactionId)
 print(f'Block included: {os.environ["EXPLORER_URL"]}/block/{blockId}')
 
 print(f'Created token: {prepared_transaction.token_id()}')