+++ sdk/examples/client/output/micro_transaction.rs
@@ -7,9 +7,10 @@
 //! However, it is possible to send a large amount and ask a slightly smaller amount in return to
 //! effectively transfer a small amount.
 //!
-//! `cargo run --example microtransaction --release`
-
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+//! Rename `.env.example` to `.env` first, then run the command:
+//! ```sh
+//! cargo run --release --example microtransaction
+//! ```
 
 use iota_sdk::{
     client::{api::GetAddressesOptions, request_funds_from_faucet, secret::SecretManager, Client, Result},
@@ -26,12 +27,11 @@ async fn main() -> Result<()> {
     // non-zero balance.
     dotenvy::dotenv().ok();
 
-    let node_url = std::env::var("NODE_URL").unwrap();
-    let explorer_url = std::env::var("EXPLORER_URL").unwrap();
-    let faucet_url = std::env::var("FAUCET_URL").unwrap();
-
-    // Create a client instance.
-    let client = Client::builder().with_node(&node_url)?.finish().await?;
+    // Create a node client.
+    let client = Client::builder()
+        .with_node(&std::env::var("NODE_URL").unwrap())?
+        .finish()
+        .await?;
 
     let secret_manager =
         SecretManager::try_from_mnemonic(std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap())?;
@@ -44,15 +44,19 @@ async fn main() -> Result<()> {
 
     let token_supply = client.get_token_supply().await?;
 
-    request_funds_from_faucet(&faucet_url, &sender_address).await?;
+    println!(
+        "Requesting funds (waiting 15s): {}",
+        request_funds_from_faucet(&std::env::var("FAUCET_URL").unwrap(), &sender_address).await?,
+    );
     tokio::time::sleep(std::time::Duration::from_secs(15)).await;
 
-    let tomorrow = (SystemTime::now() + Duration::from_secs(24 * 3600))
-        .duration_since(UNIX_EPOCH)
+    let tomorrow = (std::time::SystemTime::now() + std::time::Duration::from_secs(24 * 3600))
+        .duration_since(std::time::UNIX_EPOCH)
         .expect("clock went backwards")
         .as_secs()
         .try_into()
         .unwrap();
+
     let outputs = [
         // with storage deposit return
         BasicOutputBuilder::new_with_amount(255_100)
@@ -76,7 +80,11 @@ async fn main() -> Result<()> {
         .finish()
         .await?;
 
-    println!("Block with micro amount sent: {explorer_url}/block/{}", block.id());
+    println!(
+        "Block with micro amount sent: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
+        block.id()
+    );
     let _ = client.retry_until_included(&block.id(), None, None).await?;
 
     Ok(())