+++ sdk/examples/wallet/offline_signing/1_prepare_transaction.rs
@@ -8,19 +8,13 @@
 //! cargo run --release --all-features --example 1_prepare_transaction
 //! ```
 
-use std::env::var;
-
 use iota_sdk::{
     client::{
         api::{PreparedTransactionData, PreparedTransactionDataDto},
         constants::SHIMMER_COIN_TYPE,
         secret::SecretManager,
     },
-    wallet::{account::types::AccountAddress, ClientOptions, Result, SendAmountParams, Wallet},
-};
-use tokio::{
-    fs::File,
-    io::{AsyncReadExt, AsyncWriteExt, BufReader, BufWriter},
+    wallet::{account::types::AccountAddress, ClientOptions, Result, SendParams, Wallet},
 };
 
 const ONLINE_WALLET_DB_PATH: &str = "./examples/wallet/offline_signing/example-online-walletdb";
@@ -36,12 +30,12 @@ async fn main() -> Result<()> {
     // This example uses secrets in environment variables for simplicity which should not be done in production.
     dotenvy::dotenv().ok();
 
-    let outputs = [SendAmountParams::new(RECV_ADDRESS, SEND_AMOUNT)?];
+    let params = [SendParams::new(RECV_ADDRESS, SEND_AMOUNT)?];
 
     // Recovers addresses from example `0_address_generation`.
     let addresses = read_addresses_from_file().await?;
 
-    let client_options = ClientOptions::new().with_node(&var("NODE_URL").unwrap())?;
+    let client_options = ClientOptions::new().with_node(&std::env::var("NODE_URL").unwrap())?;
 
     // Create the wallet with the secret_manager and client options
     let wallet = Wallet::builder()
@@ -63,9 +57,9 @@ async fn main() -> Result<()> {
     // Sync the account to get the outputs for the addresses
     account.sync(None).await?;
 
-    let prepared_transaction = account.prepare_send_amount(outputs.clone(), None).await?;
+    let prepared_transaction = account.prepare_send(params.clone(), None).await?;
 
-    println!("Prepared transaction sending {outputs:?}");
+    println!("Prepared transaction sending {params:?}");
 
     write_transaction_to_file(prepared_transaction).await?;
 
@@ -73,7 +67,9 @@ async fn main() -> Result<()> {
 }
 
 async fn read_addresses_from_file() -> Result<Vec<AccountAddress>> {
-    let mut file = BufReader::new(File::open(ADDRESSES_FILE_PATH).await?);
+    use tokio::io::AsyncReadExt;
+
+    let mut file = tokio::io::BufReader::new(tokio::fs::File::open(ADDRESSES_FILE_PATH).await?);
     let mut json = String::new();
     file.read_to_string(&mut json).await?;
 
@@ -81,8 +77,10 @@ async fn read_addresses_from_file() -> Result<Vec<AccountAddress>> {
 }
 
 async fn write_transaction_to_file(prepared_transaction: PreparedTransactionData) -> Result<()> {
+    use tokio::io::AsyncWriteExt;
+
     let json = serde_json::to_string_pretty(&PreparedTransactionDataDto::from(&prepared_transaction))?;
-    let mut file = BufWriter::new(File::create(PREPARED_TRANSACTION_FILE_PATH).await?);
+    let mut file = tokio::io::BufWriter::new(tokio::fs::File::create(PREPARED_TRANSACTION_FILE_PATH).await?);
     println!("example.prepared_transaction.json:\n{json}");
     file.write_all(json.as_bytes()).await?;
     file.flush().await?;