+++ sdk/src/wallet/bindings/nodejs/types/bridge/account.ts
@@ -5,7 +5,7 @@ import type {
 } from '@iota/types';
 import type { SyncOptions, FilterOptions } from '../account';
 import type {
-    SendAmountParams,
+    SendParams,
     SendNativeTokensParams,
     SendNftParams,
     GenerateAddressOptions,
@@ -280,10 +280,10 @@ export type __PrepareOutputMethod__ = {
     };
 };
 
-export type __PrepareSendAmountMethod__ = {
-    name: 'prepareSendAmount';
+export type __PrepareSendMethod__ = {
+    name: 'prepareSend';
     data: {
-        params: SendAmountParams[];
+        params: SendParams[];
         options?: TransactionOptions;
     };
 };
@@ -320,10 +320,10 @@ export type __RetryTransactionUntilIncludedMethod__ = {
     };
 };
 
-export type __SendAmountMethod__ = {
-    name: 'sendAmount';
+export type __SendMethod__ = {
+    name: 'send';
     data: {
-        params: SendAmountParams[];
+        params: SendParams[];
         options?: TransactionOptions;
     };
 };